requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Statistical modeling (for anomaly detection)
statsmodels>=0.14.0
scikit-learn>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type checking (optional, for development)
mypy>=1.4.0

# Jupyter support
jupyter>=1.0.0
ipykernel>=6.25.0

